refactor(IN-284): 필터 영역 리팩토링
